+++ packages/block-editor/src/components/block-preview/auto.js
@@ -65,7 +65,7 @@ function AutoBlockPreview( { viewportWidth, __experimentalPadding } ) {
 					} }
 				>
 					{ contentResizeListener }
-					<MemoizedBlockList renderAppender={ false } />
+					<MemoizedBlockList appender={ null } />
 				</Iframe>
 			</Disabled>
 		</div>