+++ packages/block-editor/src/components/inner-blocks/index.js
@@ -49,6 +49,7 @@ function UncontrolledInnerBlocks( props ) {
 		__experimentalCaptureToolbars: captureToolbars,
 		__experimentalAppenderTagName,
 		renderAppender,
+		appender,
 		orientation,
 		placeholder,
 		__experimentalLayout,
@@ -91,6 +92,7 @@ function UncontrolledInnerBlocks( props ) {
 			<BlockListItems
 				rootClientId={ clientId }
 				renderAppender={ renderAppender }
+				appender={ appender }
 				__experimentalAppenderTagName={ __experimentalAppenderTagName }
 				__experimentalLayout={ __experimentalLayout }
 				wrapperRef={ wrapperRef }