+++ packages/block-library/src/group/edit.js
@@ -59,9 +59,9 @@ function GroupEdit( { attributes, setAttributes, clientId } ) {
 			: { className: 'wp-block-group__inner-container' },
 		{
 			templateLock,
-			renderAppender: hasInnerBlocks
-				? undefined
-				: InnerBlocks.ButtonBlockAppender,
+			appender: hasInnerBlocks ? undefined : (
+				<InnerBlocks.ButtonBlockAppender />
+			),
 			__experimentalLayout: layoutSupportEnabled ? usedLayout : undefined,
 		}
 	);