+++ packages/block-library/src/template-part/edit/inner-blocks.js
@@ -36,9 +36,9 @@ export default function TemplatePartInnerBlocks( {
 		value: blocks,
 		onInput,
 		onChange,
-		renderAppender: hasInnerBlocks
-			? undefined
-			: InnerBlocks.ButtonBlockAppender,
+		appender: hasInnerBlocks ? undefined : (
+			<InnerBlocks.ButtonBlockAppender />
+		),
 		__experimentalLayout: themeSupportsLayout ? usedLayout : undefined,
 	} );